Separated token tabs content to components
